==> Thesis/WORKING_FILES/02 Sampling.Rmd <==
---
title: "02 Sampling"
author: "Timur Sharifullin"
date: '2022-10-29'
output: html_document
---
# Библиотеки
```{r setup, include=FALSE}
library(tidyverse)
library(RSelenium)
library(netstat)
library(stringr)
library(dplyr)
library(vkR)
library(XML)
library(plotly)
library(rapportools)
library(beepr)
library(seleniumPipes)
library(lubridate)
library(stringi)

library(spotifyr)
Sys.setenv(SPOTIFY_CLIENT_ID = '1e1c38c8ccf64465a7490c1698dee485')
Sys.setenv(SPOTIFY_CLIENT_SECRET = 'aee21b7926b34e31abe811687f431952')

access_token <- get_spotify_access_token()
```

# Поиск трека в Spotify
Цели: 
- выгрузить даты релиза (album.release_date)
- выгрузить popularity (мало ли понадобится)
- выгрузить сингл или альбом (album.album_type)
- выгрузить артистов и ссылки на них (artists)
- выгрузить id треков (id)
```{r}
#ds_raw_names6 = data.frame(artist_def = ds_raw_names5$artist, name_def = ds_raw_names5$title, artists = NA, name = NA, id = NA, album.release_date = NA, album.album_type = NA, popularity = NA, track_url = NA, artist = NA)

for(i in seq(from = 801, to = 1118, by = 1)){
  search_query = paste(ds_raw_names5[i,1], ds_raw_names5[i,2], sep = " ")
  
  # ищем трек
  track_search = search_spotify(q = search_query,limit = 1, type = "track",
                       authorization = get_spotify_access_token(), include_meta_info = F) 
  if(nrow(track_search)!=0){
    # выбираем необходимые переменные 
    track_search = track_search %>% select(artists, name, id, album.release_date, album.album_type, popularity,
                                           track_url = external_urls.spotify)
    track_search$artist = paste(track_search[[1]][[1]][,3], collapse="; ")
    
    #переносим в таблицу
    ds_raw_names6[i,3] = track_search[1,1]
    ds_raw_names6[i,4] = track_search[1,2]
    ds_raw_names6[i,5] = track_search[1,3]
    ds_raw_names6[i,6] = track_search[1,4]
    ds_raw_names6[i,7] = track_search[1,5]
    ds_raw_names6[i,8] = track_search[1,6]
    ds_raw_names6[i,9] = track_search[1,7]
    ds_raw_names6[i,10] = track_search[1,8]
  }
}

ds_raw_names6_ = ds_raw_names6 %>% select(-artists)
write.csv(ds_raw_names6_, "C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_raw_names6.csv")
ds_raw_names6 = read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_raw_names6.csv")
ds_raw_names6 = ds_raw_names6[,-1]
```

# Отфильтровать по дате релиза 
```{r}
# избавляемся от треков, которые не в Spotify
ds_raw_names7 = na.omit(ds_raw_names6)

# Фильтруем по дате релиза
ds_raw_names7$album.release_date = ymd(ds_raw_names7$album.release_date)
ds_raw_names7 = ds_raw_names7 %>% filter(album.release_date >= "2020-07-15" & album.release_date <= "2022-03-10")
ds_raw_names7 %>% select(-artists) %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_raw_names7.csv")
```


```{r}
# создаем таблицу для всех артистов 
ds_artists_spotify = ds_raw_names7[[3]][[1]]
for (i in seq(from=2, to=nrow(ds_raw_names7), by = 1)){
  ds_raw_names7[[3]][[i]]
  ds_artists_spotify = rbind(ds_artists_spotify,ds_raw_names7[[3]][[i]])
}

# remove duplicates
ds_artists_spotify = unique(ds_artists_spotify)

# save the table
ds_artists_spotify %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artists_spotify.csv", row.names = F)
```
# Loading datasets
```{r}
ds_raw_names7 = read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_raw_names7.csv")
```

```{r}
# выгружаем жанры артистов
for (i in seq(from=201, to=442, by = 1)){
  ds_artists_spotify[i,7] = paste(get_artist(ds_artists_spotify[i,2],
                                               authorization = get_spotify_access_token())$genres, 
                                                collapse = "; ")
}

# save the table
ds_artists_spotify %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artists_spotify.csv", row.names = F)
ds_artists_spotify = read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artists_spotify.csv")

# filter artists by location in genre
ds_artists_spotify1 = ds_artists_spotify %>% mutate(is_russian = str_detect(ds_artists_spotify[,7],
                                                                           "russian|kazakh|ukrainian|belarusian|zhenskiy"))
# filter artists by the presence of cyrillic letters in the name
for (i in seq(from=1, to=nrow(ds_artists_spotify1), by = 1)){
  if(stri_detect(str = ds_artists_spotify1[i,3], regex = "[\\p{script=cyrillic}]")==T){
    ds_artists_spotify1[i,8] = TRUE
  }
}

# save the table & move to excel to manually add all the missing genres
ds_artists_spotify1 %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artists_spotify1.csv")

# load the resulting table from excel
ds_artists_spotify2 = read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artists_spotify1_fix.csv")

# filter artists by location
ds_artists_spotify2 = ds_artists_spotify2 %>% mutate(is_russian = str_detect(ds_artists_spotify2[,7],
                                                                           "russian|kazakh|ukrainian|belarusian|zhenskiy"))

# filter artists by location
ds_artists_spotify2_rus_hh = ds_artists_spotify2 %>% 
  filter(str_detect(ds_artists_spotify2[,7], "hip hop|drain|hyperpop|drill|rap|rep|grime|trap") & is_russian ==T)

ds_artists_spotify2_rus_hh$genre = ds_artists_spotify2_rus_hh[,7]
ds_artists_spotify2_rus_hh = select(ds_artists_spotify2_rus_hh, -is_russian, - V7)

# save the table
ds_artists_spotify2_rus_hh %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_artist_sample.csv", row.names = F)
```

```{r}
# FINAL TRACK SAMPLING 
ds_check = ds_raw_names7
for (i in seq(from=1, to=nrow(ds_raw_names7), by = 1)){
  ds_check[i,10] = any(unlist(strsplit(ds_raw_names7[i,9],"; ")) %in% ds_artists_spotify2_rus_hh[,3])
}
ds_music = ds_check %>% filter(ds_check[,10]==T)
ds_music = ds_music[,-10]

# fix some random duplicates
rownames(ds_music) = NULL
ds_music = ds_music[c(-259,-300),]

# save the table
ds_music %>% write.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_music.csv", row.names = F)
read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/ds_music.csv")
length(unique(unlist(strsplit(ds_music[,9],"; "))))
```
