==> Thesis/WORKING_FILES/05 all_tracks_play_count.ipynb <==
! pip install pytesseract

import keyboard as kb
import time
import pyautogui
import csv
import pandas
import re
from PIL import Image
from pytesseract import pytesseract
import numpy as np

# loading datasets
ds_music = pandas.read_csv("d.csv")
#ds_music["artists_all"][14043] = "Zomb"
#ds_music["artists_all"][14044] = "OG Buda"

# screenshots of spotify
kb.send("alt+tab")
time.sleep(0.5)
kb.send("enter")
time.sleep(0.2)
  
# create keyboard loop  
g = 0
while (g < len(ds_music)):
  search_query = " ".join([" ".join((str.split(ds_music.iloc[g][1],";"))),(ds_music.iloc[g][2])])
  kb.send("ctrl+l")
  time.sleep(1)
  kb.write(search_query)
  time.sleep(2)
  kb.send("enter")
  time.sleep(2)
  kb.send("tab")  
  
  kb.send("enter")
  time.sleep(2)
  kb.send("right")  

  myScreenshot = pyautogui.screenshot()
  path = " ".join(["C:\\Users\\timur\\OneDrive\\Documents\\R Studio Tasks\\ДИПЛОМ\\Screenshots Spotify\\",str(g), 
    re.sub(string = search_query, pattern = "[^A-Za-zА-Яа-яЁ-ё0-9\s]+", repl = ""),".png"])
  myScreenshot.save(path)
  g = g+1

kb.send("alt+tab")

ds_music

#Define path to tessaract.exe
path_to_tesseract = r'C:\Program Files\Tesseract-OCR\tesseract.exe'
#Point tessaract_cmd to tessaract.exe
pytesseract.tesseract_cmd = path_to_tesseract
#ds_music["play_counts_spotify"] = np.nan

i=0
while (i<len(ds_music)):
    #Define path to image
    search_query = " ".join([" ".join((str.split(ds_music.iloc[i][1],";"))),(ds_music.iloc[i][2])])
    path_to_image = " ".join([rf"C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/Screenshots Spotify/{str(i)}",
                              re.sub(string = search_query, pattern = "[^A-Za-zА-Яа-яЁ-ё0-9\s]+", repl = ""),".png"])
    #Open image with PIL
    img = Image.open(path_to_image)
    #Crop image
    box = (1209,626,(1209+171),(626+52))
    img2 = img.crop(box)

    box1 = (668,354,(668+526),(354+44))
    img3 = img.crop(box1)

    #Extract text from image
    text = pytesseract.image_to_string(img2)
    text1 = pytesseract.image_to_string(img3,lang="rus")

    if "слушат" in text1:
        n_play = np.NaN
    else: 
        n_play = re.sub("[^0-9]", "", str(text))

    ds_music["n_plays"][i] = n_play
    i=i+1

ds_music

ds_music.to_csv("C:\\Users\\timur\\OneDrive\\Documents\\R Studio Tasks\\ДИПЛОМ\\unique_plays1.csv")
