==> Thesis/WORKING_FILES/06 Popularity filtering.Rmd <==
---
title: "06 Popularity Filtering"
author: "Timur Sharifullin"
date: "2023-01-16"
output: html_document
---

```{r}
library(dplyr)
library(lubridate)
library(stringr)
library(spotifyr)
Sys.setenv(SPOTIFY_CLIENT_ID = '1e1c38c8ccf64465a7490c1698dee485')
Sys.setenv(SPOTIFY_CLIENT_SECRET = 'aee21b7926b34e31abe811687f431952')
```

# Загрузка Датасета
```{r setup, include=FALSE}
a = read.csv("C:\\Users\\timur\\Downloads\\Sample - Лист1 (2) (1).csv")
```

# Приведение таблицы в нормальный вид
```{r}
# переименовка столбцов
a = a %>% select(position = Pos, artist_title = Artist.and.Title, wks = Wks, t10 = T10, peak = Pk, n_appear = X.x..,
                 peak_streams = PkStreams, total_streams = Total)

# Изменение классов и пр.
a$n_appear = str_extract_all(string = a$n_appear, pattern = "([0-9])") %>% as.integer() 
a$peak_streams = str_remove_all(string = a$peak_streams, pattern = ",") %>% as.integer()
a$total_streams = str_remove_all(string = a$total_streams, pattern = ",") %>% as.integer()
```

# Добавляем нужные столбцы в таблицу
```{r}
a = mutate(a, artist.name = NA, artist.id = NA,
           track.name = NA, track.id = NA, track.url = NA, 
           album.name = NA, album.release_date = NA, album.album_type = NA,
           duration_ms = NA)
```

# Поиск трека в Spotify
```{r}
for(i in seq(from = 1801, to = 1927, by = 1)){
  search_query = a[i,2]
  
  # ищем трек
  track_search = search_spotify(q = search_query, 
                                limit = 1, 
                                type = "track",
                       authorization = get_spotify_access_token(), include_meta_info = F)

  if(nrow(track_search) != 0){
    # выбираем необходимые переменные 
    track_search = track_search %>% select(artists, track.name = name, track.id = id, track.url = external_urls.spotify,
                                           album.name, album.release_date, album.album_type, 
                                           duration_ms)
    track_search$artist.name = paste(track_search[[1]][[1]][,3], collapse="; ")
    track_search$artist.id = paste(track_search[[1]][[1]][,2], collapse="; ")
    
    #переносим в таблицу
    a$artist.name[i] = track_search$artist.name[1]
    a$artist.id[i] = track_search$artist.id[1]
    a$track.name[i] = track_search$track.name[1]
    a$track.id[i] = track_search$track.id[1]
    a$track.url[i] = track_search$track.url[1]
    a$album.name[i] = track_search$album.name[1]
    a$album.release_date[i] = track_search$album.release_date[1]
    a$album.album_type[i] = track_search$album.album_type[1]
    a$duration_ms[i] = track_search$duration_ms[1]
  }
}
```

# Отфильтровать по дате релиза 
```{r}
a1 = a

# Фильтруем по дате релиза
a1$album.release_date = ymd(a1$album.release_date)
a1 = a1 %>% filter(album.release_date >= "2020-07-15" & album.release_date <= "2022-03-10")
```
# Оставить только уникальные значения
```{r}
keep= NA
for(g in seq(from = 1, to = nrow(a1), by = 1)){
  keep[g]=!(isTRUE(a1[g,2] %in% c(a1[1:(g-1),2])))
}
# Первая строчка всегда обозначается как FALSE, поэтому фиксим
keep[1] = TRUE
# Оставляем только нужные значения в таблице
a2 = a1[keep,]
```

# Жанры артистов
```{r}
# Список уникальных исполнителей
artist_to_genre = unlist(strsplit(a2$artist.id, "; ")) %>% unique() %>% as.data.frame() %>% mutate(genre = NA)

# выгружаем жанры артистов
for (i in seq(from=401, to=586, by = 1)){
  artist_to_genre[i,2] = paste(get_artist(artist_to_genre[i,1],authorization = get_spotify_access_token())$genres, 
                                                collapse = "; ")
}
```

# Добавляем в эту таблицу имена
```{r}
aid = unlist(strsplit(a2$artist.id, "; "))
aname = unlist(strsplit(a2$artist.name, "; "))
to_genre_again = data.frame(aid, aname) %>% unique

artist_to_genre = artist_to_genre %>% rename(aid = ".")

genre_artists = merge(artist_to_genre, to_genre_again, by = 'aid')
genre_artists = genre_artists %>% select(aname, genre, aid)
```

# Добавляем некоторые жанры вручную
```{r}
write.csv(genre_artists, file="C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/genre_artists.csv", row.names = F)
genre_artists1 = read.csv("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/genre_artists1.csv")
```

# Фильтруем по жанру
```{r}
# filter artists by location in genre
genre_artists2 = filter(genre_artists1, str_detect(genre_artists1[,2],"russian|kazakh|ukrainian|belarusian|zhenskiy")==T)

# filter artists by genre
genre_artists3 = genre_artists2 %>% 
  filter(str_detect(genre_artists2[,2], "hip hop|drain|hyperpop|drill|rap|rep|grime|trap"))
```

# Фильтруем основную таблицу
```{r}
a3 = a2 
for (i in seq(from=1, to=nrow(a2), by = 1)){
  a3[i,18] = any(unlist(strsplit(a2[i,10],"; ")) %in% genre_artists3[,3])
}
ds_music = a3 %>% filter(a3[,18]==T)
ds_music = ds_music[,-18]
```

# Выгружаем таблицу
```{r}
write.csv(ds_music, file="C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/df_music_new1801.csv", row.names = F)
```
