==> Thesis/WORKING_FILES/13 Finalizing.Rmd <==
---
title: "13 Finalizing Table"
author: "Timur Sharifullin"
date: "2023-04-09"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
library(prettydoc)
library(dplyr)
library(stringr)
library(ggplot2)
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
t0 = read.csv("new_combine_features_full_audio_final_0.csv")
t1 = read.csv("new_combine_features_full_audio_final_1.csv")
t2 = read.csv("new_combine_features_full_audio_final_2.csv")
t31 = read.csv("new_combine_features_full_audio_final_31.csv")

t01 = rbind(t0,t1)
t23 = rbind(t2,t31)
t = rbind(t01,t23)
```

```{r}
t = t[,-1]
row.names(t) = NULL

t$track_id = str_sub(t$song_name, start = 0, end = 22)
t11 = t %>% select(-song_name)
row.names(t11) = NULL
t11_new = t11[!duplicated(t11$track_id),]
```

```{r}
fin = read.csv("ds_fin_donwload_links.csv")

matching_rows = fin$track_id %in% t11_new$track_id
fin1 = fin[matching_rows,]
```

```{r}
fin_ds = merge(fin1,t11_new, by = "track_id")
summary(fin_ds)
```


```{r}
fin_ds[15294:15295,7] = 1
fin_ds[15294:15295,11] = "A#"
fin_ds[15294:15295,12] = "minor"
fin_ds[15294:15295,13] = "A# minor"

fin_ds[413:414,7] = 1

fin_ds[7806:7807,7] = 4
fin_ds[7806:7807,11] = "G"
fin_ds[7806:7807,12] = "major"
fin_ds[7806:7807,13] = "G major"

fin_ds[7247,7] = 1
fin_ds[7247,11] = "C#"
fin_ds[7247,12] = "major"
fin_ds[7247,13] = "C# major"

fin_ds[7789,7] = 1
fin_ds[7247,11] = "F"
fin_ds[7247,12] = "minor"
fin_ds[7247,13] = "F minor"

fin_ds = fin_ds[-6170,]
row.names(fin_ds) = NULL

fin_ds1 = fin_ds %>% unique()
```

```{r}
row.names(fin_ds1) = NULL
summary(fin_ds1)
```

```{r}
write.csv(fin_ds1,"fin_ds1.csv",row.names = F)
```


# Arranging
```{r}
fin_ds1 = fin_ds1 %>% arrange(artist_name,album_release_date,track_number)
```

# Preprocessing for mutating hit_n
```{r}
df <- fin_ds1

df <- df %>%
  group_by(artist_name) %>%
  mutate(occur = cumsum(X == "hit"))
df$hit_n = NA
df
```

# Mutate Column with hit_n
```{r}
# Split the data frame by artist_name
df_list <- split(df, df$artist_name)

# Loop through each subset of the data frame
for (i in 1:length(df_list)) {
  
  # Loop through the rows of the current subset
  for (j in 1:nrow(df_list[[i]])) {
    
    cat(paste("Artist №", i, "|","Track №", j),"\r")
    
    # condition for the first track (always = 0)
    if(j==1){
      df_list[[i]]$hit_n[j] = 0
      next
    }
        # condition for the second track 
    ## if release date of previous track is not similar to release date of current track
    if(j!=1 & df_list[[i]]$album_release_date[j] != df_list[[i]]$album_release_date[j-1]){
      
      df_list[[i]]$hit_n[j] = df_list[[i]]$occur[j-1]
      next
    }

    ## if release date of previous track is similar to release date of current track
    if(j!=1 & df_list[[i]]$album_release_date[j] == df_list[[i]]$album_release_date[j-1]){
      
      df_list[[i]]$hit_n[j] = df_list[[i]]$hit_n[j-1]
      next
    }
  }
}
```


# Final Table with hit_n
```{r}
df_devil = do.call(rbind, df_list)
df_devil = df_devil %>% select(-occur)
```


## Editing the track_status
```{r}
# copy df
df_devil1 = df_devil

# create track_status
df_devil1$track_status = paste0(df_devil1$X, "_", df_devil1$hit_n)


# editing track_status - plus 0 before 1-digit numbers for correct sorting
for(i in seq(1,nrow(df_devil1),1)){
  if(df_devil1$X[i] == "hit" & nchar(df_devil1$track_status[i])==5){
    df_devil1$track_status[i] =paste0(str_split(df_devil1$track_status[i],"_", simplify = TRUE)[1],
                                      0,
                                      str_split(df_devil1$track_status[i],"_", simplify = TRUE)[2])
  }
  
  if(df_devil1$X[i] == "none" & nchar(df_devil1$track_status[i])==6){
    
    df_devil1$track_status[i] = paste0(str_split(df_devil1$track_status[i],"_", simplify = TRUE)[1],
                                       0,
                                       str_split(df_devil1$track_status[i],"_", simplify = TRUE)[2])
  }
}

# removing "_" from track status
df_devil1$track_status = str_remove_all(df_devil1$track_status,"_") %>% as.factor()
```


```{r}
write.csv(df_devil1, "C:\\Users\\timur\\OneDrive\\Documents\\R Studio Tasks\\ДИПЛОМ\\FINAL_TABLE.csv",row.names = F)
```
