==> Thesis/WORKING_FILES/15 PCA.Rmd <==
---
title: "15 PCA"
author: "Timur Sharifullin"
date: "2023-04-11"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,message= F, warning=F)

library(dplyr)
library(stringr)
library(ggplot2)
library(lubridate)
library(plotly)
library(FactoMineR)
library(factoextra)
library(ggbiplot)
library(htmltools)
library("tuneR")
setWavPlayer(shQuote("C:/Program Files (x86)/AIMP/AIMP.exe"))

df_devil1 = read.csv("FINAL_TABLE.csv")
df_devil1 = dplyr::rename(df_devil1, rmsP_median = rmseP_median)
df_devil1 = dplyr::rename(df_devil1, rmsP_std = rmseP_std)
df_devil1 = dplyr::rename(df_devil1, rmsH_median = rmseH_median)
df_devil1 = dplyr::rename(df_devil1, rmsH_std = rmseH_std)
df_devil1$artist_name = as.factor(df_devil1$artist_name)
df_devil1$X = as.factor(df_devil1$X)
df_devil1$album_release_date = ymd(df_devil1$album_release_date)
df_devil1$key_name = as.factor(df_devil1$key_name) 
df_devil1$mode_name = as.factor(df_devil1$mode_name) 
df_devil1$key_mode = as.factor(df_devil1$key_mode) 
df_devil1$remake = as.factor(df_devil1$remake)
df_devil1$track_status = as.factor(df_devil1$track_status)
final_df = df_devil1
```

```{r}
final_df$file_link = NA 
for(i in seq(1,nrow(final_df),1)){
  file_link = paste("C:/Users/timur/OneDrive/Documents/R Studio Tasks/ДИПЛОМ/Audio New/",
                                 final_df$track_id[i],
                                 " - ",
                                 final_df$artist_name[i],
                                 " - ",
                                 final_df$track_name[i], ".mp3", sep = "")
  final_df$file_link[i] = gsub('"|/|\\?|\\*|!', '', file_link, fixed = TRUE)
}
```


## Audio Characteristics 

At the moment - these characteristics have only a general description - that is, they will not necessarily mean exactly that, since the interpretation often depends on the genre of music - that is, you need to conduct PCA and consider individual tracks in order to most correctly interpret the data variables

1. The higher is *bw_std*, the more complex in frequency-terms is the sound
2. The higher is *centroid_median*, the brighter is the sound
3. The higher is *centroid_std*, the more variability in sound and therefore - more complexity
4. The higher is *contrast_median*, the more sharpness in sound
5. The higher is *polyfeat_median* faster tempo or more energetic beat 
6. The higher is *polyfeat_std*, the more complex or varied rhythmic or melodic patterns
7. The higher is *rmsH_median*, the more melodic and tonal elements in the music, such as instrumentals, vocals, or synthesized sounds
8. The higher is *rmsH_std*, the loudness of the harmonic content changes frequently 
9. The higher is *rmsP_median*, the more intense and energetic beat
10. The higher is *rmsP_std*, the more dynamic and varied beat
11. The higher is *zcr_median*, the greater presence of percussive elements


```{r}
library(ggcorrplot)
corr_matrix = cor(final_df[,18:31])
cp = ggcorrplot(corr_matrix,hc.order = TRUE, outline.col = "black")
ggplotly(cp)
```

SENT TO PCA: 

1. zcr_median & centroid_median (0.81)
The higher is *zcr_median*, the greater presence of percussive elements
The higher is *centroid_median*, the brighter is the sound

2. polyfeat_std & rmsH_std (0.81)
The higher is *polyfeat_std*, the more complex or varied rhythmic or melodic patterns (A low standard deviation indicates that the spectral content of the signal is relatively uniform or balanced, while a high standard deviation indicates that the spectral content is more varied or unbalanced.)
The higher is *rmsH_std*, the loudness of the harmonic content changes frequently

3. rmsH_median & polyfeat_median (-0.73)
The higher is *polyfeat_median* faster tempo or more energetic beat 
The higher is *rmsH_median*, the more melodic and tonal elements in the music, such as instrumentals, vocals, or synthesized sounds

4. bw_std & centroid_std (0.6) - complexity
The higher is *bw_std*, the more complex in frequency-terms is the sound
The higher is *centroid_std*, the more variability in sound and therefore - more complexity

```{r}
names = paste(final_df$artist_name, "-", final_df$track_nam)
links = final_df$file_link
```


### 1 zcr_median & centroid_median (0.81)

```{r}
pca1 = prcomp(final_df[,c("zcr_median","centroid_median")], center = T, scale = T)
summary(pca1)
```

```{r}
options(scipen=999)
pca1$sdev ^ 2
```

91% variance - PC1
9% variance - PC2

```{r}
pca1$rotation
```

Zero crossings allow for a rough estimation of dominant frequency and the spectral centroid [https://www.sciencedirect.com/topics/engineering/zero-crossing-rate]


```{r}
ggplotly(ggbiplot::ggbiplot(pca1,repel=T, alpha = 0.6,labels = names, labels.size = 2))
#, repel = T
```


```{r}
dtpca1 = data.frame(tname = names, pc1 = pca1$x[,1], pc2 = pca1$x[,2], link = links)
dtpca1 = dtpca1 %>% arrange(desc(pc1))
dtpca1
```

```{r}
audio_play = readMP3(dtpca1$link[11600])
play(audio_play) 
```

LOW PC1
Niman - INTERNATIONAL BODY (Ambient, No Voice, Much Low-Freq and Mid)
Seryoga - Послание из космоса (Voice + Ambient, Mostly Mid freq)
Pika - China (Instrumental, Mostly Low + Mid)
вышел покурить - завтра (Monotonic Rap, slow+reverb type beat, no high freq)
Pretty Scream - Штанги (Instrumental, Calm, Lack of High-Freq)
Oxxxymiron - Стихи о неизвестном солдате (из трибьюта Осипу Мандельштаму «Сохрани мою речь навсегда») (Low-freq Monotonic Poetics)
Pica - CHina (Bassy instrumental, no hihats, lack of high freq)
SP4K - Гипнагоджик Поп (Bassy instrumental, lack of high freq)
Face - Смерть (Bassy-Mid Soundfield)
I61 - Night Heavy (Leonel Remix) (Hypno Bass)

HIGH PC1 

GENJUTSU - Mode Shadowraze (Aggressive, Noisy, Boosted)
Slava Marlow - Рок Звезда (Keedfour Remix) [Bonus Track] (Mid Freq, very bright electronic sound, lack of vocals, lack of bass)
Naitivihod - 8 bit death (VERY NOISY, Bad Mixing, Electronic Music)
gitarakuru - tears (Noisy, Hyperpop, very bright)
PLOHOYPAREN - Волк (Squeaky Hyperpop Sound)
MUKKA - ТВНБ (Bad Mixing, No low-freq, guitar, vocals)

The lower is PC1 - the more bass in music, the higher is PC1 - the more high-frequency instruments (synts or high-pitched vocals or distortion) or noise is in music

PC1 - Sound Brightness/Noisiness

Low PC2
KURT92 - INTRO (No music, only high-freq scratches and glitches)
25/17 - Майор (Only Voice)
MUKKA - ТВНБ (Bad Mixing, No low-freq, guitar, vocals)
Slava Marlow - Рок Звезда (Keedfour Remix) [Bonus Track] - (Mid Freq, very bright electronic sound, lack of vocals, lack of bass)
LIZER - Послушай (Only Voice)
Naitivihod - 8 bit death (VERY NOISY, Bad Mixing, Electronic Music)
gitarakuru - Heavy Rain (Bad Mixing, Music, Lack of Bass)
Naitivihod - спасибо, что приютила (Voice, Poetics)
Horus - Inferno (instrumental, metal)

High PC2 
twikipedia - logout  (Отрывистый Hyperpop, высокие частоты, Squeaky Vocals)
BUSHIDO ZHO - Баиньк (Detroit Rap, much variable percussion)
SEEYASIDE - 2021 (Not much percussion, Much High-Freq, like LUVV66)
ROSS - Движ Париж (Bright Hihats)
BOOKER - Нить (Bright Hihats)

PC2 - hard to interpret.


## 2 polyfeat_std & rmsH_std (0.81) 

```{r}
pca2 = prcomp(final_df[,c("polyfeat_std","rmsH_std")], center = T, scale = T)
summary(pca2)
```

```{r}
pca2$sdev ^ 2
```

90% variance - PC1
10% variance - PC2

```{r}
pca2$rotation
```

The higher is *polyfeat_std*, the more complex or varied rhythmic or melodic patterns. A low standard deviation indicates that the spectral content of the signal is relatively uniform or balanced, while a high standard deviation indicates that the spectral content is more varied or unbalanced. The standard deviation of the first-order polynomial coefficients can be interpreted as a measure of the spectral balance or tonality of the signal. In modern rap music, the tonality is often characterized by heavily processed and distorted sounds, such as distorted 808 basslines and aggressive vocal effects. A higher standard deviation of the coefficients may indicate a more varied and complex spectral content, reflecting the diverse range of sounds and textures in modern rap music.

The higher is *rmsH_std*, the loudness of the harmonic content changes frequently

```{r}


dtpca2 = data.frame(tname = names, pc1 = pca2$x[,1], pc2 = pca2$x[,2], link = links)
dtpca2 = dtpca2 %>% arrange(pc2)
dtpca2
```

```{r}
audio_play = readMP3(dtpca2$link[17])
play(audio_play) 
```

```{r}
dtpca2$tname[2]
```


HIGHEST PC1 (>3) - Only voice
HIGHER 1 PC1 - chill / lyrical sound

LOWER PC1 - music with frequently changing drum/tonal patterns + hard bass distorted sound, more bounce
LOWEST PC1 - more experimental loud distorted bass sound with super hard hits, not like a bounce, but more like shakes

------------------------------------
The lower the PC1 - the more Hard-hitting, Gritty or Aggressive is track, the higher the PC1 - the more Laid-back, Groovy or Lyrical is track.
------------------------------------

PC2 is hard to interpret.

## 3 rmsH_median & polyfeat_median (-0.73)

```{r}
pca3 = prcomp(final_df[,c("rmsH_median","polyfeat_median")], center = T, scale = T)
summary(pca3)
```

```{r}
pca3$sdev ^ 2
```

87% variance - PC1
13% variance - PC2

```{r}
pca3$rotation
```

```{r}
ggplotly(ggbiplot::ggbiplot(pca3,repel=T, alpha = 0.6,labels = names, labels.size = 2))

dtpca3 = data.frame(tname = names, pc1 = pca3$x[,1], pc2 = pca3$x[,2], link = links)
dtpca3 = dtpca3 %>% arrange(pc2)
dtpca3
```

```{r}
audio_play = readMP3(dtpca3$link[14610])
play(audio_play) 
```

 
*polyfeat_median*: In modern rap music, the median value of the first-order polynomial coefficients can be interpreted as an estimate of the overall loudness of the signal. The genre is known for its emphasis on powerful, bass-heavy beats and vocals that are often mixed loud and upfront in the mix. Therefore, a higher median value of the coefficients may indicate a louder and more impactful sound.

The higher is *rmsH_median* (Root Mean Square of Harmonic Part of the song), the more (the louder) melodic and tonal elements in the music, such as instrumentals, vocals, or synthesized sounds

Tracks that have lower values on PC1 being louder and less melodic, and tracks with higher values being quieter and more melodic.

PC2 - is hard to interpret.

## 4 bw_std & centroid_std (0.6)

```{r}
pca4 = prcomp(final_df[,c("bw_std","centroid_std")], center = T, scale = T)
summary(pca4)
```

```{r}
pca4$sdev ^ 2
```

80% variance - PC1
20% variance - PC2

```{r}
ggplotly(ggbiplot::ggbiplot(pca4,repel=T, alpha = 0.6,labels = names, labels.size = 2))

pca4$rotation
summary(pca4)
fviz_pca_biplot(pca4, alpha.ind = 0)
```

```{r}
dtpca4 = data.frame(tname = names, pc1 = pca4$x[,1], pc2 = pca4$x[,2], link = links)
dtpca4 = dtpca4 %>% arrange(desc(pc2))
dtpca4
```

```{r}
audio_play = readMP3(dtpca4$link[104])
play(audio_play) 
```

A higher *spectral centroid standard deviation* may indicate a wider range of frequencies and a brighter, more "open" sound. This could be indicative of a track with more high-frequency elements, such as bright synths, hi-hats, or vocal harmonies. In contrast, a lower spectral centroid standard deviation may indicate a narrower range of frequencies and a darker, more "closed" sound. This could be indicative of a track with more low-frequency elements, such as basslines, kick drums, or deep vocal timbres.

A higher *standard deviation of spectral bandwidth* in a rap song can indicate that the song has more varied and diverse frequency content, with a wider range of both high and low frequencies present. This could be due to the use of a diverse range of instruments and sounds, or to more complex mixing and production techniques that involve filtering and manipulating different frequency ranges.

PC1 - one-way directed (negative) 
Так, получается, что треки с меньшим PC1 имеют высокие "bw_std" и "centroid_std" - wider range of both high and low frequencies present & more "open" sound. This could be indicative of a track with more high-frequency elements, such as bright synths, hi-hats, or vocal harmonies.

PC1: The higher the value of PC1, the higher the spectral complexity and variability of the music. This can be interpreted as the music having a more diverse range of frequencies and tones, with a wider distribution of spectral energy. The negative rotation values for both spectral centroid standard deviation and spectral bandwidth standard deviation indicate that higher values of these features contribute to a lower score on PC1. In the context of modern rap music, this may suggest that tracks with lower PC1 scores may be more simplistic and repetitive in terms of their overall sound.

PC1 = Overall Spectral Variability. This component captures the degree to which the song's spectral characteristics vary across different frequencies. A higher PC1 score suggests a more dynamic, diverse, and complex soundscape in the music, with more pronounced changes in tone and timbre.

PC2 is hard to interpret.


## FINALLY
```{r}
final_df$sound_brightness_noisiness = pca1$x[,1]
final_df$laid_back_groovy_lyrical = pca2$x[,1]
final_df$quiet_and_melodic = pca3$x[,1]
final_df$overall_spectral_Variability = pca4$x[,1]

final_df1 = final_df %>% select(-zcr_median,-centroid_median,-polyfeat_std,-rmsH_std,-rmsH_median,-polyfeat_median,
                                -bw_std,-centroid_std)
```



```{r}
library(ggcorrplot)
corr_matrix = cor(final_df1[,c(18:23,27:30)])
cp = ggcorrplot(corr_matrix,hc.order = TRUE, outline.col = "black")
ggplotly(cp)

final_df1[,c(18:23,27:30)] = scale(final_df1[,c(18:23,27:30)],center = T)
```

## 0 hits

```{r}
final_df1_0 = final_df1 %>% filter(hit_n ==0)

model <- glm(X ~ rmsP_median + rmsP_std + contrast_median + contrast_std + tonnetz_median + tonnetz_std +sound_brightness_noisiness +
               laid_back_groovy_lyrical + quiet_and_melodic + overall_spectral_Variability,
            data = final_df1_0, family = "binomial")

# summarize the model results
#summary(model)

predicted_probs <- predict(model, type = "response")

# compare predictions to actual outcomes
library(pROC)  # load package for ROC analysis
roc_obj <- roc(final_df1_0$X, predicted_probs)
auc <- auc(roc_obj)  # calculate AUC of ROC curve
accuracy <- mean((predicted_probs >= 0.5) == final_df1_0$X)

cat("Model_0 accuracy:", accuracy," | ", "AUC:", auc)
```



```{r}
library(car)
vif(model)
```
