==> Thesis/WORKING_FILES/17 Logistic Regression.Rmd <==
---
title: "16 Logistic Regression"
author: "Timur Sharifullin"
date: "2023-04-12"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
library(dplyr)
library(stringr)
library(ggplot2)
library(lubridate)
library(pscl)
knitr::opts_chunk$set(echo = TRUE)
final_df = read.csv("fin_df_distance.csv")
final_df = final_df[,-1]
final_df$X = as.factor(final_df$X) 
final_df1 = final_df[,c(5, 18, 21:30)]
```

```{r}
fin  = final_df %>% filter(hit_n == 20) %>% group_by(artist_name) %>% summarize(n = n()) %>% arrange(n)


as.data.frame(summary(as.factor(fin$n)))
```
0 - 282 
1 - 238
2 - 170
3 - 140
4 - 115
5 - 96
6 - 73
7 - 64
8 - 54
9 - 48
10 - 50
11 - 43
12 - 36

```{r}
c1 = c(282,238,170,140,115,96,73,64,54,48,50,43,36)
c2 = c(rep(282,13))
round(c1/c2,2)*100
```


```{r}
m667p = glm(formula = X ~ hit_n, data = final_df,  family=binomial(link = "probit"))
summary(m667p)
pR2(m667p)['McFadden']
```


```{r}
final_df_orig = read.csv("final_df1.csv") 
final_df_orig = final_df_orig[,c(5,24, 18:23, 27:30)]
final_df_orig$X = as.factor(final_df_orig$X)
final_df1 = final_df_orig
```


## Dividing dataset to parts

```{r}
ds0 = final_df1 %>% filter(hit_n == 0) %>% select(-hit_n)
ds1 = final_df1 %>% filter(hit_n == 1) %>% select(-hit_n)
ds2 = final_df1 %>% filter(hit_n == 2) %>% select(-hit_n)
ds3 = final_df1 %>% filter(hit_n == 3) %>% select(-hit_n)
ds4 = final_df1 %>% filter(hit_n == 4) %>% select(-hit_n)
ds5 = final_df1 %>% filter(hit_n == 5) %>% select(-hit_n)
ds6 = final_df1 %>% filter(hit_n == 6) %>% select(-hit_n)
ds7 = final_df1 %>% filter(hit_n == 7) %>% select(-hit_n)
ds8 = final_df1 %>% filter(hit_n == 8) %>% select(-hit_n)
ds9 = final_df1 %>% filter(hit_n == 9) %>% select(-hit_n)
ds10 = final_df1 %>% filter(hit_n == 10) %>% select(-hit_n)
```

#```{r}
library(ggcorrplot)
library(plotly)
corr_matrix = cor(final_df[,21:30])
cp = ggcorrplot(corr_matrix,hc.order = TRUE, outline.col = "black")
ggplotly(cp)
#```


## Первый хит и до него

```{r}
final_df1 %>% 
  ggplot(aes(x=X, fill=X)) +
  geom_bar() +
  geom_text(aes(label = ..count..), stat = "count", vjust = 1.5, colour = "white")
```

```{r}
m666p = glm(formula = X ~ ., data = final_df1,  family=binomial(link = "probit"))
summary(m666p)
pR2(m666p)['McFadden']
```


```{r}
library(ROSE)
ds0.rose <- ROSE(X ~ ., data = ds0, seed = 1)$data
ds1.rose <- ROSE(X ~ ., data = ds1, seed = 1)$data
ds2.rose <- ROSE(X ~ ., data = ds2, seed = 1)$data
ds3.rose <- ROSE(X ~ ., data = ds3, seed = 1)$data
ds4.rose <- ROSE(X ~ ., data = ds4, seed = 1)$data
ds5.rose <- ROSE(X ~ ., data = ds5, seed = 1)$data
ds6.rose <- ROSE(X ~ ., data = ds6, seed = 1)$data
ds7.rose <- ROSE(X ~ ., data = ds7, seed = 1)$data
ds8.rose <- ROSE(X ~ ., data = ds8, seed = 1)$data
ds9.rose <- ROSE(X ~ ., data = ds9, seed = 1)$data
ds10.rose <- ROSE(X ~ ., data = ds10, seed = 1)$data
```


```{r}
library(sjPlot)
plot_model(m0p, type = "pred", terms = c("X", "contrast_std"))
```

```{r}
library(caret)
mydata_balanced <- downSample(x = ds0[, -ncol(ds0)], y = ds0$X)

ggplot(ds0, aes(x = contrast_std, y = X)) +
  geom_point() +
  geom_smooth(method = "glm", method.args = list(family = "binomial"), se = FALSE)
```


```{r}

```


```{r}
m0p = glm(formula = X ~ ., data = ds0,  family=binomial(link = "probit"))

sum_m0 = data.frame(variable = row.names(as.data.frame(summary(m0p)$coefficients[,1])),
                    estimates = summary(m0p)$coefficients[,1],
                    p_value = summary(m0p)$coefficients[,4],
                    signif = if_else(summary(m0p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m0p)['McFadden']))

#---------------------------------------------------------------------------------------

m1p = glm(formula = X ~ ., data = ds1,  family=binomial(link = "probit"))

sum_m1 = data.frame(variable = row.names(as.data.frame(summary(m1p)$coefficients[,1])),
                    estimates = summary(m1p)$coefficients[,1],
                    p_value = summary(m1p)$coefficients[,4],
                    signif = if_else(summary(m1p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m1p)['McFadden']))

#---------------------------------------------------------------------------------------

m2p = glm(formula = X ~ ., data = ds2,  family=binomial(link = "probit"))

sum_m2 = data.frame(variable = row.names(as.data.frame(summary(m2p)$coefficients[,1])),
                    estimates = summary(m2p)$coefficients[,1],
                    p_value = summary(m2p)$coefficients[,4],
                    signif = if_else(summary(m2p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m2p)['McFadden']))

#---------------------------------------------------------------------------------------

m3p = glm(formula = X ~ ., data = ds3,  family=binomial(link = "probit"))

sum_m3 = data.frame(variable = row.names(as.data.frame(summary(m3p)$coefficients[,1])),
                    estimates = summary(m3p)$coefficients[,1],
                    p_value = summary(m3p)$coefficients[,4],
                    signif = if_else(summary(m3p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m3p)['McFadden']))
#---------------------------------------------------------------------------------------

m4p = glm(formula = X ~ ., data = ds4,  family=binomial(link = "probit"))

sum_m4 = data.frame(variable = row.names(as.data.frame(summary(m4p)$coefficients[,1])),
                    estimates = summary(m4p)$coefficients[,1],
                    p_value = summary(m4p)$coefficients[,4],
                    signif = if_else(summary(m4p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m4p)['McFadden']))
#---------------------------------------------------------------------------------------

m5p = glm(formula = X ~ ., data = ds5,  family=binomial(link = "probit"))

sum_m5 = data.frame(variable = row.names(as.data.frame(summary(m5p)$coefficients[,1])),
                    estimates = summary(m5p)$coefficients[,1],
                    p_value = summary(m5p)$coefficients[,4],
                    signif = if_else(summary(m5p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m5p)['McFadden']))
#---------------------------------------------------------------------------------------

m6p = glm(formula = X ~ ., data = ds6,  family=binomial(link = "probit"))

sum_m6 = data.frame(variable = row.names(as.data.frame(summary(m6p)$coefficients[,1])),
                    estimates = summary(m6p)$coefficients[,1],
                    p_value = summary(m6p)$coefficients[,4],
                    signif = if_else(summary(m6p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m6p)['McFadden']))
#---------------------------------------------------------------------------------------

m7p = glm(formula = X ~ ., data = ds7,  family=binomial(link = "probit"))

sum_m7 = data.frame(variable = row.names(as.data.frame(summary(m7p)$coefficients[,1])),
                    estimates = summary(m7p)$coefficients[,1],
                    p_value = summary(m7p)$coefficients[,4],
                    signif = if_else(summary(m7p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m7p)['McFadden']))
#---------------------------------------------------------------------------------------

m8p = glm(formula = X ~ ., data = ds8,  family=binomial(link = "probit"))

sum_m8 = data.frame(variable = row.names(as.data.frame(summary(m8p)$coefficients[,1])),
                    estimates = summary(m8p)$coefficients[,1],
                    p_value = summary(m8p)$coefficients[,4],
                    signif = if_else(summary(m8p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m8p)['McFadden']))
#---------------------------------------------------------------------------------------

m9p = glm(formula = X ~ ., data = ds9,  family=binomial(link = "probit"))

sum_m9 = data.frame(variable = row.names(as.data.frame(summary(m9p)$coefficients[,1])),
                    estimates = summary(m9p)$coefficients[,1],
                    p_value = summary(m9p)$coefficients[,4],
                    signif = if_else(summary(m9p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m9p)['McFadden']))
#---------------------------------------------------------------------------------------

m10p = glm(formula = X ~ ., data = ds10,  family=binomial(link = "probit"))

sum_m10 = data.frame(variable = row.names(as.data.frame(summary(m10p)$coefficients[,1])),
                    estimates = summary(m10p)$coefficients[,1],
                    p_value = summary(m10p)$coefficients[,4],
                    signif = if_else(summary(m10p)$coefficients[,4]<0.05, "true",""),
                    R2 = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,pR2(m10p)['McFadden']))
```

```{r}
sum_m0[1,1] = "Model 0"
sum_m1[1,1] = "Model 1"
sum_m2[1,1] = "Model 2"
sum_m3[1,1] = "Model 3"
sum_m4[1,1] = "Model 4"
sum_m5[1,1] = "Model 5"
sum_m6[1,1] = "Model 6"
sum_m7[1,1] = "Model 7"
sum_m8[1,1] = "Model 8"
sum_m9[1,1] = "Model 9"
sum_m10[1,1] = "Model 10"

sum_m_similarity = rbind(sum_m0,sum_m1,sum_m2,
                     sum_m3,sum_m4,sum_m5,
                     sum_m6,sum_m7,sum_m8,
                     sum_m9,sum_m10)

write.csv(sum_m_similarity, "sum_m_without_balance.csv",row.names = F)
```


```{r}
library(sjPlot)
tab_model(m0p,
          m1p,
          m2p,
          m3p,
          m4p,
          m5p,
          m6p,
          m7p,
          m8p,
          m9p,
          m10p,
          show.aic = T, show.loglik = T,
          dv.labels = c('m0',"m1","m2","m3","m4","m5",
                        'm6',"m7","m8","m9","m10"))
```
