==> Thesis/WORKING_FILES/21 Hypothesis 2_with_log_reg NEW.Rmd <==
---
title: "21 Hypothesis 2_with_log_reg"
author: "Timur Sharifullin"
date: "2023-04-26"
output: html_document
---

```{r}
library(ggplot2)
library(dplyr)
library(tidyverse)
library(see)
```


```{r}
final_df = read.csv("fin_df_distance_year.csv")
final_df$X = factor(final_df$X, levels = c("none","hit"))

#final_df1 = final_df[,c(5, 21:30)] %>% unique()

final_df1 = final_df[,c(3,5,18,21:30)] 

colnames(final_df1)[4:13] = c("drum_character", "drum_complexity", "number_of_instruments",
                        "mixing_quality", "harmonic_richness", "REMOVE",
                        "mixing_character", "emotional_character", "REMOVE",
                        "REMOVE")

final_df1 = final_df1[,-c(9,12,13)]
final_df1
```

```{r}

```


# Rule Of Thumb

```{r}
table(final_df1$X,final_df1$hit_n)
ds = final_df1 %>% filter(hit_n < 6) %>% dplyr::select(-hit_n) %>% unique()
```


# Subsets

```{r}
set.seed(2023)

summary(ds$X)

data = as.data.frame(table(ds$artist_name)) %>% arrange(-Freq)
data

data1 = data %>% filter(Freq>=summary(data$Freq)[5]) %>% dplyr::select(Var1)
data1 = as.character(data1$Var1)
vec = c()
for(i in seq(1,nrow(ds),1)){
  if(as.character(ds$artist_name[i]) %in% data1){
    vec = append(vec,i)
  }
}

ds1 = ds[-vec,]
summary(ds1$X)
ds1 = ds1 %>% dplyr::select(-artist_name)
```




# Distributions

```{r}
for (i in 2:8) {
  p =ggplot(data = ds1) +
    geom_histogram(aes(x = ds1[, i],fill = X), bins=1000) +
    labs(x = colnames(ds1)[i])
    
  # Print each plot
  print(p)
}
```






#  Univariates Analysis

## Without interaction terms

```{r}
colnames(ds1)
```


```{r}
library(mgcv)
u1 = gam(X~s(drum_character),data = ds1, family = binomial)
u2 = gam(X~s(drum_complexity),data = ds1, family = binomial)
u3 = gam(X~s(number_of_instruments),data = ds1, family = binomial)

u4 = gam(X~s(mixing_quality),data = ds1, family = binomial)
u5 = gam(X~s(harmonic_richness),data = ds1, family = binomial)
u6 = gam(X~s(mixing_character),data = ds1, family = binomial)

u7 = gam(X~s(emotional_character),data = ds1, family = binomial)


summary(u1)$s.pv

plot(ggeffects::ggpredict(u1))
summary(u2)$s.pv
plot(ggeffects::ggpredict(u2))
summary(u3)$s.pv
plot(ggeffects::ggpredict(u3))


summary(u4)$s.pv
plot(ggeffects::ggpredict(u4))
summary(u5)$s.pv
plot(ggeffects::ggpredict(u5))
summary(u6)$s.pv
plot(ggeffects::ggpredict(u6))

summary(u7)$s.pv
plot(ggeffects::ggpredict(u7))
```

rmsP_median - 0.05312208 .
rmsP_std - 0.1398791 (non-significant)
contrast_median - 0.000255627 *** 
contrast_std - 0.04607081 * 
tonnetz_median - 0.0195016 * 
tonnetz_std - 0.03323094 *
sound_brightness_noisiness - 0.3912911 (non-significant)
laid_back_groovy_lyrical -  0.0002602455 *** 
quiet_and_melodic - 0.7443353 (non-significant)
overall_spectral_Variability - 0.6995442 (non-significant)


-------------------
(Just audio features)

rmsP_median - 0.02204191 +
rmsP_std - 6.620711e-05 +
contrast_median - 0.004538091 +
contrast_std - 0.9921364 +
tonnetz_median - 0.4699638 
tonnetz_std - 0.2881367
sound_brightness_noisiness - 0.9334526
laid_back_groovy_lyrical -  0.09611018
quiet_and_melodic - 0.4204202
overall_spectral_Variability - 0.4897683

## Without interaction terms
```{r}
gam_1 = gam(X ~
              s(contrast_median) +
              s(laid_back_groovy_lyrical),
            data = ds1,
            family = binomial)

summary(gam_1)
plot(ggeffects::ggpredict(gam_1))
gam.check(gam_1)
```


tonnetz_median - same 
laid_back_groovy_lyrical

```{r}
ds = final_df1 %>% filter(hit_n < 6) %>% unique()

gam_1 = gam(X ~
              ti(rmsP_median) +
              ti(rmsP_std) +
              ti(contrast_median) +
              ti(tonnetz_median) +
              ti(sound_brightness_noisiness) +
              ti(laid_back_groovy_lyrical) +
              ti(quiet_and_melodic) +
              ti(overall_spectral_Variability),
            data = ds,
            family = binomial)

summary(gam_1)
```


## DS1
```{r}
u1 = gam(X~s(rmsP_median),data = ds1, family = binomial)
u2 = gam(X~s(rmsP_std),data = ds1, family = binomial)
u3 = gam(X~s(contrast_median),data = ds1, family = binomial)

u4 = gam(X~s(contrast_std),data = ds1, family = binomial)
u5 = gam(X~s(tonnetz_median),data = ds1, family = binomial)
u6 = gam(X~s(tonnetz_std),data = ds1, family = binomial)

u7 = gam(X~s(sound_brightness_noisiness),data = ds1, family = binomial)
u8 = gam(X~s(laid_back_groovy_lyrical),data = ds1, family = binomial)
u9 = gam(X~s(quiet_and_melodic),data = ds1, family = binomial)

u10 = gam(X~s(overall_spectral_Variability),data = ds1, family = binomial)

summary(u1)
plot(ggeffects::ggpredict(u1))
summary(u2)
plot(ggeffects::ggpredict(u2))
summary(u3)
plot(ggeffects::ggpredict(u3))


summary(u4)
plot(ggeffects::ggpredict(u4))
summary(u5)
plot(ggeffects::ggpredict(u5))
summary(u6)
plot(ggeffects::ggpredict(u6))

summary(u7)
plot(ggeffects::ggpredict(u7))
summary(u8)
plot(ggeffects::ggpredict(u8))
summary(u9)
plot(ggeffects::ggpredict(u9))

summary(u10)
plot(ggeffects::ggpredict(u10))
```

rmsP_std - U-Shape
sound_brightness_noisiness - downslope shape
laid_back_groovy_lyrical - kinda U-shape
quiet_and_melodic - downslope shape

```{r}
m10 = gam(X~s(rmsP_std) + s(sound_brightness_noisiness) + s(laid_back_groovy_lyrical) + s(quiet_and_melodic),
    data = ds1, family = binomial)

summary(m10)
```

## DS2
```{r}
u1 = gam(X~s(rmsP_median),data = ds2, family = binomial)
u2 = gam(X~s(rmsP_std),data = ds2, family = binomial)
u3 = gam(X~s(contrast_median),data = ds2, family = binomial)

u4 = gam(X~s(contrast_std),data = ds2, family = binomial)
u5 = gam(X~s(tonnetz_median),data = ds2, family = binomial)
u6 = gam(X~s(tonnetz_std),data = ds2, family = binomial)

u7 = gam(X~s(sound_brightness_noisiness),data = ds2, family = binomial)
u8 = gam(X~s(laid_back_groovy_lyrical),data = ds2, family = binomial)
u9 = gam(X~s(quiet_and_melodic),data = ds2, family = binomial)

u10 = gam(X~s(overall_spectral_Variability),data = ds2, family = binomial)

summary(u1)
plot(ggeffects::ggpredict(u1))
summary(u2)
plot(ggeffects::ggpredict(u2))
summary(u3)
plot(ggeffects::ggpredict(u3))


summary(u4)
plot(ggeffects::ggpredict(u4))
summary(u5)
plot(ggeffects::ggpredict(u5))
summary(u6)
plot(ggeffects::ggpredict(u6))

summary(u7)
plot(ggeffects::ggpredict(u7))
summary(u8)
plot(ggeffects::ggpredict(u8))
summary(u9)
plot(ggeffects::ggpredict(u9))

summary(u10)
plot(ggeffects::ggpredict(u10))
```

No significant variables

laid_back_groovy_lyrical


## DS3
```{r}
u1 = gam(X~s(rmsP_median),data = ds3, family = binomial)
u2 = gam(X~s(rmsP_std),data = ds3, family = binomial)
u3 = gam(X~s(contrast_median),data = ds3, family = binomial)

u4 = gam(X~s(contrast_std),data = ds3, family = binomial)
u5 = gam(X~s(tonnetz_median),data = ds3, family = binomial)
u6 = gam(X~s(tonnetz_std),data = ds3, family = binomial)

u7 = gam(X~s(sound_brightness_noisiness),data = ds3, family = binomial)
u8 = gam(X~s(laid_back_groovy_lyrical),data = ds3, family = binomial)
u9 = gam(X~s(quiet_and_melodic),data = ds3, family = binomial)

u10 = gam(X~s(overall_spectral_Variability),data = ds3, family = binomial)

summary(u1)
plot(ggeffects::ggpredict(u1))
summary(u2)
plot(ggeffects::ggpredict(u2))
summary(u3)
plot(ggeffects::ggpredict(u3))


summary(u4)
plot(ggeffects::ggpredict(u4))
summary(u5)
plot(ggeffects::ggpredict(u5))
summary(u6)
plot(ggeffects::ggpredict(u6))

summary(u7)
plot(ggeffects::ggpredict(u7))
summary(u8)
plot(ggeffects::ggpredict(u8))
summary(u9)
plot(ggeffects::ggpredict(u9))

summary(u10)
plot(ggeffects::ggpredict(u10))
```

No signinficant variables

## DS4
```{r}
u1 = gam(X~s(rmsP_median),data = ds4, family = binomial)
u2 = gam(X~s(rmsP_std),data = ds4, family = binomial)
u3 = gam(X~s(contrast_median),data = ds4, family = binomial)

u4 = gam(X~s(contrast_std),data = ds4, family = binomial)
u5 = gam(X~s(tonnetz_median),data = ds4, family = binomial)
u6 = gam(X~s(tonnetz_std),data = ds4, family = binomial)

u7 = gam(X~s(sound_brightness_noisiness),data = ds4, family = binomial)
u8 = gam(X~s(laid_back_groovy_lyrical),data = ds4, family = binomial)
u9 = gam(X~s(quiet_and_melodic),data = ds4, family = binomial)

u10 = gam(X~s(overall_spectral_Variability),data = ds4, family = binomial)

summary(u1)
plot(ggeffects::ggpredict(u1))
summary(u2)
plot(ggeffects::ggpredict(u2))
summary(u3)
plot(ggeffects::ggpredict(u3))


summary(u4)
plot(ggeffects::ggpredict(u4))
summary(u5)
plot(ggeffects::ggpredict(u5))
summary(u6)
plot(ggeffects::ggpredict(u6))

summary(u7)
plot(ggeffects::ggpredict(u7))
summary(u8)
plot(ggeffects::ggpredict(u8))
summary(u9)
plot(ggeffects::ggpredict(u9))

summary(u10)
plot(ggeffects::ggpredict(u10))
```

tonnetz_std
laid_back_groovy_lyrical  

## DS5
```{r}
u1 = gam(X~s(rmsP_median),data = ds5, family = binomial)
u2 = gam(X~s(rmsP_std),data = ds5, family = binomial)
u3 = gam(X~s(contrast_median),data = ds5, family = binomial)

u4 = gam(X~s(contrast_std),data = ds5, family = binomial)
u5 = gam(X~s(tonnetz_median),data = ds5, family = binomial)
u6 = gam(X~s(tonnetz_std),data = ds5, family = binomial)

u7 = gam(X~s(sound_brightness_noisiness),data = ds5, family = binomial)
u8 = gam(X~s(laid_back_groovy_lyrical),data = ds5, family = binomial)
u9 = gam(X~s(quiet_and_melodic),data = ds5, family = binomial)

u10 = gam(X~s(overall_spectral_Variability),data = ds5, family = binomial)

summary(u1)
plot(ggeffects::ggpredict(u1))
summary(u2)
plot(ggeffects::ggpredict(u2))
summary(u3)
plot(ggeffects::ggpredict(u3))


summary(u4)
plot(ggeffects::ggpredict(u4))
summary(u5)
plot(ggeffects::ggpredict(u5))
summary(u6)
plot(ggeffects::ggpredict(u6))

summary(u7)
plot(ggeffects::ggpredict(u7))
summary(u8)
plot(ggeffects::ggpredict(u8))
summary(u9)
plot(ggeffects::ggpredict(u9))

summary(u10)
plot(ggeffects::ggpredict(u10))
```

rmsP_std





```{r}
gam.model <- gam(X ~ 
              rmsP_median*hit_n +
              s(rmsP_std)*hit_n +
              s(contrast_median)*hit_n +
              s(contrast_std)*hit_n +
              s(tonnetz_median)*hit_n +
              s(tonnetz_std)*hit_n +
              s(sound_brightness_noisiness)*hit_n +
              s(quiet_and_melodic)*hit_n +
              s(laid_back_groovy_lyrical)*hit_n +
              s(overall_spectral_Variability)*hit_n,
            data = final_df1,
            family = binomial,
            select = TRUE)
summary(gam.model)


gam.model <- gam(X ~ 
              rmsP_median*hit_n +
              rmsP_std*hit_n,
            data = final_df1,
            family = binomial)

#R-sq.(adj) =  0.00793   Deviance explained =  1.1%


#plot(gam.model)
#plot(ggeffects::ggpredict(gam.model), facets = TRUE) + see::scale_color_flat()

plot(ggeffects::ggpredict(gam.model))
```

1. U-Shape
2. U-Shape
3. Linear (down)
4. U-shape
5. Curved down (Slightly Reverse U-shaped)
6. Linear
7. No relationship
8. Linear (down)
9. S-shape
10. U-shpe + curved(down)


```{r}
gam.model_1o <- gam(X ~ rmsP_median +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_2o <- gam(X ~ s(rmsP_median) +
              rmsP_std +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_3o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              contrast_median +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_4o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              contrast_std +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_5o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              tonnetz_median +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_6o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              tonnetz_std +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_7o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              sound_brightness_noisiness +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_8o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              quiet_and_melodic +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_9o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              laid_back_groovy_lyrical +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)

gam.model_10o <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              overall_spectral_Variability,
            data = final_df1,
            family = binomial)


anova(gam.model_1o,
      gam.model_2o,
      gam.model_3o,
      gam.model_4o,
      gam.model_5o,
      gam.model_6o,
      gam.model_7o,
      gam.model_8o,
      gam.model_9o,
      gam.model_10o,
      gam.model,
      test="Chisq")
```


```{r}
gam.model_31 <- gam(X ~ s(rmsP_median) +
              rmsP_std +
              contrast_median +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = final_df1,
            family = binomial)
summary(gam.model_31)
```
















## Разделение на subsets
```{r}
set.seed(2023)

ds0 = final_df1 %>% filter(hit_n == 0) %>% select(-hit_n) %>% unique()
ds1 = final_df1 %>% filter(hit_n == 1) %>% select(-hit_n) %>% unique()
ds2 = final_df1 %>% filter(hit_n == 2) %>% select(-hit_n) %>% unique()
ds3 = final_df1 %>% filter(hit_n == 3) %>% select(-hit_n) %>% unique()
ds4 = final_df1 %>% filter(hit_n == 4) %>% select(-hit_n) %>% unique()
ds5 = final_df1 %>% filter(hit_n == 5) %>% select(-hit_n) %>% unique()

summary(ds0$X)
summary(ds1$X)
summary(ds2$X)
summary(ds3$X)
summary(ds4$X)
summary(ds5$X)
```

# DS0

```{r}
gam0 <- gam(X ~ s(rmsP_median) + s(rmsP_std) + s(contrast_median) + s(contrast_std) +
                              s(tonnetz_median) + s(tonnetz_std) + s(sound_brightness_noisiness) +
                              s(quiet_and_melodic) + s(laid_back_groovy_lyrical) + s(overall_spectral_Variability),
                            data = ds0, 
                            family = binomial)
summary(ds0$X)
summary(gam0)

plot(ggeffects::ggpredict(gam0), facets = TRUE) + see::scale_color_flat()
plot(gam0)
```

None of the variables are significant.

# DS1

```{r}

gam1 <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = ds1,
            family = binomial,
            select = TRUE)


summary(gam1)


plot(ggeffects::ggpredict(gam1), facets = TRUE) + see::scale_color_flat()
```

```{r}
model_ts <- gam(X ~ 
               s(rmsP_median,bs="ts") +
               s(rmsP_std,bs="ts") +
               s(contrast_median,bs="ts") +
               s(contrast_std,bs="ts") +
               s(tonnetz_median,bs="ts") +
               s(tonnetz_std,bs="ts") +
               s(sound_brightness_noisiness,bs="ts") +
               s(quiet_and_melodic,bs="ts") +
               s(laid_back_groovy_lyrical,bs="ts") +
               s(overall_spectral_Variability,bs="ts"),
             data = ds1,
             family = binomial)

model_cr <- gam(X ~ 
               s(rmsP_median,bs="cr") +
               s(rmsP_std,bs="cr") +
               s(contrast_median,bs="cr") +
               s(contrast_std,bs="cr") +
               s(tonnetz_median,bs="cr") +
               s(tonnetz_std,bs="cr") +
               s(sound_brightness_noisiness,bs="cr") +
               s(quiet_and_melodic,bs="cr") +
               s(laid_back_groovy_lyrical,bs="cr") +
               s(overall_spectral_Variability,bs="cr"),
             data = ds1,
             family = binomial)

model_cc <- gam(X ~ 
               s(rmsP_median,bs="cc") +
               s(rmsP_std,bs="cc") +
               s(contrast_median,bs="cc") +
               s(contrast_std,bs="cc") +
               s(tonnetz_median,bs="cc") +
               s(tonnetz_std,bs="cc") +
               s(sound_brightness_noisiness,bs="cc") +
               s(quiet_and_melodic,bs="cc") +
               s(laid_back_groovy_lyrical,bs="cc") +
               s(overall_spectral_Variability,bs="cc"),
             data = ds1,
             family = binomial)

model_tp <- gam(X ~ 
               s(rmsP_median,bs="tp") +
               s(rmsP_std,bs="tp") +
               s(contrast_median,bs="tp") +
               s(contrast_std,bs="tp") +
               s(tonnetz_median,bs="tp") +
               s(tonnetz_std,bs="tp") +
               s(sound_brightness_noisiness,bs="tp") +
               s(quiet_and_melodic,bs="tp") +
               s(laid_back_groovy_lyrical,bs="tp") +
               s(overall_spectral_Variability,bs="tp"),
             data = ds1,
             family = binomial)

summary(model_ts)
summary(model_cr)
summary(model_cc)
summary(model_tp)
?s()
```



contrast_median & tonnetz_std have slightly U-shape relationship with dependent variable, however, these variable are not significant according to the model.

rmsP_std, contrast_std, laid_back_groove_lyrical,  are curved to the left, so the more is the difference (novelty) of the song in comparison to the non-hits. 

# DS2

```{r}
gam2 <- gam(X ~ s(rmsP_median) + s(rmsP_std) + s(contrast_median) + s(contrast_std) +
                              s(tonnetz_median) + s(tonnetz_std) + s(sound_brightness_noisiness) +
                              s(quiet_and_melodic) + s(laid_back_groovy_lyrical) + s(overall_spectral_Variability),
                            data = ds2,
                            family = binomial)

summary(gam2)

library(see)
plot(ggeffects::ggpredict(gam2), facets = TRUE) + see::scale_color_flat()
```

# DS3

```{r}
gam3 <- gam(X ~ s(rmsP_median) + s(rmsP_std) + s(contrast_median) + s(contrast_std) +
                              s(tonnetz_median) + s(tonnetz_std) + s(sound_brightness_noisiness) +
                              s(quiet_and_melodic) + s(laid_back_groovy_lyrical) + s(overall_spectral_Variability),
                            data = ds3,
                            family = binomial)

summary(gam3)

library(see)
plot(ggeffects::ggpredict(gam3), facets = TRUE) + see::scale_color_flat()
```

# DS4

```{r}
gam4 <- gam(X ~ s(rmsP_median) + s(rmsP_std) + s(contrast_median) + s(contrast_std) +
                              s(tonnetz_median) + s(tonnetz_std) + s(sound_brightness_noisiness) +
                              s(quiet_and_melodic) + s(laid_back_groovy_lyrical) + s(overall_spectral_Variability),
                            data = ds4,
                            family = binomial)

summary(gam4)

library(see)
plot(ggeffects::ggpredict(gam4), facets = TRUE) + see::scale_color_flat()
```

# DS5

```{r}
gam5 <- gam(X ~ s(rmsP_median) + s(rmsP_std) + s(contrast_median) + s(contrast_std) +
                              s(tonnetz_median) + s(tonnetz_std) + s(sound_brightness_noisiness) +
                              s(quiet_and_melodic) + s(laid_back_groovy_lyrical) + s(overall_spectral_Variability),
                            data = ds5,
                            family = binomial)

summary(gam5)

library(see)
plot(ggeffects::ggpredict(gam5), facets = TRUE) + see::scale_color_flat()
```

```{r}
library(rcompanion)
compareGLM(model0,
           model1,
           model2,
           model3,
           model4,
           model5)
```




```{r}
dat = ds5

gam.model <- gam(X ~ s(rmsP_median) +
              s(rmsP_std) +
              s(contrast_median) +
              s(contrast_std) +
              s(tonnetz_median) +
              s(tonnetz_std) +
              s(sound_brightness_noisiness) +
              s(quiet_and_melodic) +
              s(laid_back_groovy_lyrical) +
              s(overall_spectral_Variability),
            data = dat,
            family = binomial,
            select = TRUE)


summary(gam.model)


plot(ggeffects::ggpredict(gam.model), facets = TRUE) + see::scale_color_flat()
```
